requirements.txt
asgiref==3.8.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==4.2.20
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.7.0
firebase-admin==6.6.0
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
mysqlclient==2.2.6
pip==25.0.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pymongo==3.11.4
pyparsing==3.2.1
pytz==2025.1
requests==2.32.3
rsa==4.9
setuptools==75.9.1
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
gunicorn==23.0.0
